Moved getKanjis() logic into its own file.
