+++ app.js
@@ -2,6 +2,7 @@ import * as helper from './js/helper.js';
 import * as exampleSentences from './js/example_sentences.js';
 import * as grammars from './js/grammars.js';
 import * as vocabs from './js/vocabs.js';
+import * as kanjis from './js/kanjis.js';
 
 import * as error from 'console';
 import * as puppeteer from 'puppeteer';
@@ -26,12 +27,12 @@ const jsonDirPath =  "jsons/"
 // const mainVocab = "WPJLPT-N5-1";
 // const currentVocab = mainVocab;
 
-const mainKanji = "買";
-const multiMeaningKanji = "外";
-const varPlusLessUsedSubKanji = "酒";
-const kanjiComponent = "录";
-const currentKanji = varPlusLessUsedSubKanji;
-const currentKanjis = [mainKanji, multiMeaningKanji, varPlusLessUsedSubKanji, kanjiComponent, "伸", "近", "差", "録"];
+// const mainKanji = "買";
+// const multiMeaningKanji = "外";
+// const varPlusLessUsedSubKanji = "酒";
+// const kanjiComponent = "录";
+// const currentKanji = varPlusLessUsedSubKanji;
+// const currentKanjis = [mainKanji, multiMeaningKanji, varPlusLessUsedSubKanji, kanjiComponent, "伸", "近", "差", "録"];
 
 (async () => {
     //Launch puppeteer
@@ -54,16 +55,16 @@ const currentKanjis = [mainKanji, multiMeaningKanji, varPlusLessUsedSubKanji, ka
         //const vocabPageObj = await vocabs.getVocabPageFromBrowser(browser, currentVocab, cookies);
         
         //Get kanji object
-        const kanjiObj = await getKanjis(browser, currentKanjis, cookies);
+        //const kanjiObj = await kanjis.getKanjis(browser, currentKanjis, cookies);
 
 
         //Display run time (ms)
         console.log(Date.now() - beforePageCallsTime);
 
         //Convert example sentence to JSON
-        const kanjiJson = JSON.stringify(kanjiObj, null, 4);
+        //const kanjiJson = JSON.stringify(kanjiObj, null, 4);
         //Save Sentence JSON to file
-        await helper.saveDataToFile(kanjiJson, jsonDirPath + "test5.json");
+        //await helper.saveDataToFile(kanjiJson, jsonDirPath + "test6.json");
 
         //Take a screenshot of the current page
             //await page.screenshot({path: screenshotsDirPath + "page.png", fullPage: true});
@@ -83,94 +84,3 @@ const currentKanjis = [mainKanji, multiMeaningKanji, varPlusLessUsedSubKanji, ka
     
 })();
 
-async function getKanjis(browser, kanjiUrlSuffixes, cookies) {
-    //Map url suffixes to getKanjiDetail() Promises
-    let pages = [];//To close later
-    const kanjiPromises = kanjiUrlSuffixes.map(async (kanjiUrlSuffix) => {
-        //Open new page
-        const page = await browser.newPage();
-        //Store page to close later
-        pages.push(page);
-        //Get kanji
-        try {
-            //Load cookies
-            await page.setCookie(...cookies);
-            //Get kanji
-            const kanjiObj = await getKanjiDetail(page, kanjiUrlSuffix);
-            //Return
-            return kanjiObj;
-        } catch (error) {
-            console.error('Error in getKanjis():', error);
-            return false;
-        }
-    });
-    //Process results
-    try {
-        const kanjis = await Promise.all(kanjiPromises);
-        //Add list to object
-        const kanjisObj = { kanjis: kanjis };
-        //Return
-        return kanjisObj;
-    } catch(error) {
-        console.error('Error in getKanjis():', error);
-        return false;
-    } finally {
-        //Close all pages
-        await Promise.all(pages.map(page => page.close()));
-    }
-}
-
-async function getKanjiDetail(page, kanjiUrlSuffix) {
-    //Base url
-    const baseUrl = 'https://www.kanshudo.com/kanji/';
-    //Full url
-    const kanjiUrl = baseUrl + kanjiUrlSuffix;
-    //Get kanji detail
-    try {
-        //Navigate to page
-        await page.goto(kanjiUrl);
-        //Get kanji components in parallel
-        const [kanjiIds, kanjiDef, kanjiType, componentIds] = await Promise.all([
-            helper.getIds(page, '#main-content .bodyarea .kanjirow.level0 .kr_container .kanji', 'id', '^(k_kan_)'),
-            getKanjiDetailDefinition(page),
-            getKanjiType(page),
-            helper.getIds(page, '#main-content .bodyarea .kanjirow.level1 .kr_container .kanji', 'id', '^(k_kan_)')
-        ])
-        //Grab the kanji id (only 1 id per page)
-        const kanjiId = kanjiIds[0];
-        //Return
-        return {
-            id: kanjiId,
-            kanji: kanjiDef.kanji,
-            meaning: kanjiDef.meaning,
-            type: kanjiType,
-            component_ids: componentIds
-        };
-    } catch (error) {
-        console.error('Error in getKanjiDetail():', error);
-        return false;
-    }
-}
-
-async function getKanjiType(page) {
-    return await page.evaluate(() => {
-        return document.querySelector('#main-content .bodyarea .kdetailsrow .kdetails2 div .typemessage').textContent;
-    });
-}
-
-async function getKanjiDetailDefinition(page) {
-    //Get the kanji definition sentence (includes kanji + filler text + definition)
-    const defWithFiller = await page.evaluate(() => {
-        return document.querySelector('#main-content .bodyarea h1').textContent;
-    });
-    //Split the definition by "mean" (kanji means 'def')
-    const kanjiDefSplit = defWithFiller.split('means');
-    //Get each component of the split
-    const kanji = kanjiDefSplit[0].trim();
-    const kanjiDef = kanjiDefSplit[1].trim().replaceAll("'", ''); //Remove surrounding ''
-    //Return
-    return {
-        kanji: kanji,
-        meaning: kanjiDef
-    }
-}
