--- jsons/test1.json
@@ -1,3 +1,0 @@
-{
-    "id": "111"
-}
