--- jsons/test2.json
@@ -1,5 +1,0 @@
-{
-    "id": "831",
-    "kanji": "買",
-    "meaning": "buy"
-}
