--- jsons/test3.json
@@ -1,8 +1,0 @@
-{
-    "id": "831",
-    "kanji": "買",
-    "meaning": "buy",
-    "sub_kanji_ids": [
-        "54"
-    ]
-}
