--- jsons/test4.json
@@ -1,10 +1,0 @@
-{
-    "id": "1428",
-    "kanji": "酒",
-    "meaning": "sake",
-    "type": "Jōyō (常用)",
-    "component_ids": [
-        "4473",
-        "1427"
-    ]
-}
